Рефакторинг сервисов - создание хелпера NetworkServiceHelper
