+++ FakeNFT/Services/ProfileService.swift
@@ -1,16 +1,17 @@
 import Foundation
 
 final class ProfileService {
-    static let shared = ProfileService(networkClient: DefaultNetworkClient())
-    private let networkClient: NetworkClient
+    static let shared = ProfileService(networkHelper: NetworkServiceHelper(networkClient: DefaultNetworkClient()))
     
-    init(networkClient: NetworkClient) {
-        self.networkClient = networkClient
+    private let networkHelper: NetworkServiceHelper
+    
+    init(networkHelper: NetworkServiceHelper) {
+        self.networkHelper = networkHelper
     }
     
-    func fetchProfile(request: NetworkRequest = FetchProfileNetworkRequest(),
-                      completion: @escaping (Result<UserProfile, Error>) -> Void) {
-        networkClient.send(request: request, type: UserProfile.self) { result in
+    func fetchProfile(completion: @escaping (Result<UserProfile, Error>) -> Void) {
+        let request = FetchProfileNetworkRequest()
+        networkHelper.fetchData(request: request, type: UserProfile.self) { result in
             switch result {
             case .success(let profile):
                 completion(.success(profile))
@@ -23,7 +24,7 @@ final class ProfileService {
     func updateProfile(with userProfileModel: UserProfile,
                        completion: @escaping (Result<UserProfile, Error>) -> Void) {
         let request = UpdateProfileNetworkRequest(userProfile: userProfileModel)
-        networkClient.send(request: request, type: UserProfile.self) { result in
+        networkHelper.fetchData(request: request, type: UserProfile.self) { result in
             switch result {
             case .success(let updatedProfile):
                 completion(.success(updatedProfile))
@@ -32,5 +33,9 @@ final class ProfileService {
                 completion(.failure(error))
             }
         }
+    }
+    
+    func stopAllTasks() {
+        networkHelper.stopAllTasks()
     }
 }
